requirements.txt
alabaster==0.7.13
asttokens==2.4.1
attrs==23.2.0
Babel==2.14.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
codecov==2.1.13
contourpy==1.2.0
coverage==7.4.0
cycler==0.12.1
decorator==5.1.1
docutils==0.20.1
exceptiongroup==1.2.0
executing==2.0.1
flake8==6.1.0
flake8-annotations==3.0.1
flake8-import-order==0.18.2
Flask==3.0.0
fonttools==4.47.2
furl==2.1.3
grpcio==1.60.0
grpcio-tools==1.60.0
idna==3.6
imagesize==1.4.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
numpy==1.26.3
orderedmultidict==1.0.1
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pluggy==1.3.0
prompt-toolkit==3.0.43
protobuf==4.25.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.17.2
pyparsing==3.1.1
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.0
typing_extensions==4.9.0
urllib3==1.26.7
wcwidth==0.2.12
Werkzeug==3.0.1
zipp==3.17.0
